requirements.txt
Cartopy==0.23.0
inference==0.12.0
matplotlib==3.9.0
netCDF4==1.7.0
netCDF4==1.7.0
numpy==1.24.4
pandas==2.2.2
Pillow==10.3.0
roboflow==1.1.32
streamlit==1.29.0
supervision==0.21.0
xarray==2024.6.0
~treamlit==1.35.0
